[MQTT5] First pass property & compilation clean-up
